# apps/ade-web/public/openapi/ade-api-openapi.json
{
  "components": {
    "schemas": {
...
      "ApiRuntimeToolListResponse": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiRuntimeToolResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "search": {
            "title": "Search",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "search",
          "limit",
          "items"
        ],
        "title": "ApiRuntimeToolListResponse",
        "type": "object"
      },
...
      "ApiToolCenterItemResponse": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "last_updated_at": {
            "default": "",
            "title": "Last Updated At",
            "type": "string"
          },
          "managed": {
            "title": "Managed",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "read_only": {
            "title": "Read Only",
            "type": "boolean"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "source_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Code"
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "tool_id": {
            "title": "Tool Id",
            "type": "string"
          },
          "tool_type": {
            "title": "Tool Type",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "tool_id",
          "name",
          "description",
          "tool_type",
          "source_type",
          "managed",
          "read_only",
          "archived"
        ],
        "title": "ApiToolCenterItemResponse",
        "type": "object"
      },
      "ApiToolCenterListResponse": {
        "properties": {
          "include_archived": {
            "title": "Include Archived",
            "type": "boolean"
          },
          "include_builtin": {
            "title": "Include Builtin",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiToolCenterItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "include_archived",
          "include_builtin",
          "items"
        ],
        "title": "ApiToolCenterListResponse",
        "type": "object"
      },
      "ApiToolTestInvokeResponse": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "expected_tool_matched": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Tool Matched"
          },
          "expected_tool_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Tool Name"
          },
          "input": {
            "title": "Input",
            "type": "string"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "tool_call_count": {
            "title": "Tool Call Count",
            "type": "integer"
          },
          "tool_return_count": {
            "title": "Tool Return Count",
            "type": "integer"
          }
        },
        "required": [
          "agent_id",
          "input",
          "tool_call_count",
          "tool_return_count",
          "result"
        ],
        "title": "ApiToolTestInvokeResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ToolCenterCreateRequest": {
        "properties": {
          "default_requires_approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Requires Approval"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "enable_parallel_execution": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Parallel Execution"
          },
          "npm_requirements": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Npm Requirements"
          },
          "pip_requirements": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pip Requirements"
          },
          "return_char_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Char Limit"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "source_code": {
            "title": "Source Code",
            "type": "string"
          },
          "source_type": {
            "default": "python",
            "title": "Source Type",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "slug",
          "source_code"
        ],
        "title": "ToolCenterCreateRequest",
        "type": "object"
      },
      "ToolCenterUpdateRequest": {
        "properties": {
          "default_requires_approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Requires Approval"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "enable_parallel_execution": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Parallel Execution"
          },
          "npm_requirements": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Npm Requirements"
          },
          "pip_requirements": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pip Requirements"
          },
          "return_char_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Char Limit"
          },
          "source_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Code"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "ToolCenterUpdateRequest",
        "type": "object"
      },
      "ToolTestInvokeRequest": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "expected_tool_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Tool Name"
          },
          "input": {
            "title": "Input",
            "type": "string"
          },
          "override_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Model"
          },
          "override_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override System"
          },
          "retry_count": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Count"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 600.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "agent_id",
          "input"
        ],
        "title": "ToolTestInvokeRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Provides the versioned Agent Studio, lab, content-center, catalog, and test APIs used by ADE Web, workflows, and first-class developer clients.",
    "summary": "Feature-aligned runtime and authoring APIs for Letta Open ADE",
    "title": "ADE API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/tool-center/invocations": {
      "post": {
        "operationId": "api_invoke_tool_probe_api_v2_tool_center_invocations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolTestInvokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolTestInvokeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invoke a runtime message to validate tool-call behavior",
        "tags": [
          "Tool Center"
        ]
      }
    },
    "/api/v2/tool-center/runtime-tools": {
      "get": {
        "operationId": "api_list_runtime_tools_api_v2_tool_center_runtime_tools_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "title": "Search",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiRuntimeToolListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List tools for Toolbench discovery",
        "tags": [
          "Tool Center"
        ]
      }
    },
    "/api/v2/tool-center/tools": {
      "get": {
        "operationId": "api_tool_center_list_tools_api_v2_tool_center_tools_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_builtin",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Builtin",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_source",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Source",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "title": "Search",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolCenterListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Tool Center entries",
        "tags": [
          "Tool Center"
        ]
      },
      "post": {
        "operationId": "create_tool_api_v2_tool_center_tools_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCenterCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolCenterItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create managed custom tool",
        "tags": [
          "Tool Center"
        ]
      }
    },
    "/api/v2/tool-center/tools/{slug}": {
      "get": {
        "operationId": "api_tool_center_get_tool_api_v2_tool_center_tools__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_source",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Source",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolCenterItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Tool Center managed custom tool",
        "tags": [
          "Tool Center"
        ]
      },
      "patch": {
        "operationId": "update_tool_api_v2_tool_center_tools__slug__patch",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCenterUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolCenterItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update managed custom tool",
        "tags": [
          "Tool Center"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local ADE API",
      "url": "http://127.0.0.1:8000"
    }
  ],
...
}